Extract reading/loading of listings and items into standalone file
